Fix int8 pixel-index overflow corrupting periodic-box parity
The periodic branch of create_graph stored the direction-pixel index
with int(..., int8), but node%phi is int16 and pixel indices reach
ntheta*nphi = 256 on the default 8x32 grid: anything above 127 wrapped
to a zero/negative index, dir_x/y/z were read out of bounds (no bounds
checks under -Ofast), and every -box -4pcfp chirality sign was
orientation-dependent garbage -- an all-left-handed test catalogue
gives a parity-odd fraction of ~0.01 instead of ~1. The non-periodic
pass was updated to int16 in the v2.5.0 16-bit pixel bump; the periodic
branch was missed, and survived because the historical 4x16 = 64-pixel
grid fit in int8 and no test covered periodic parity. -exactparity was
unaffected (no pixels involved).

Add a regression test (test_periodic_parity): randomly rotated chiral
tetrahedra strictly interior to the box (so minimum-image == direct
distances) must give NNNN and NNNN_odd identical between the -box and
non-periodic runs to file precision, and a near-unity parity-odd
fraction. Random rotations matter: a fixed orientation touches only a
few pixels and can miss the wrap by luck. Verified the test fails on
the pre-fix code with exactly the periodic-vs-direct NNNN_odd mismatch
(and that with both branches broken the equality alone would not catch
it, hence the additional coherence assertion).

+++ src/graph_module.F90
@@ -183,7 +183,10 @@ contains
             i_theta = min(cfg%n_theta_dir, max(1, int(floor(theta_sph * cfg%n_theta_dir / PI)) + 1))
             i_phi   = min(cfg%n_phi_dir,   max(1, int(floor((phi_sph + PI) * cfg%n_phi_dir / (2.0d0*PI))) + 1))
 
-            output(i)%phi(k) = int((i_theta - 1) * cfg%n_phi_dir + i_phi, int8)
+            ! int16, matching node%phi and the non-periodic pass: pixel
+            ! indices reach n_theta*n_phi (256 on the default 8x32 grid),
+            ! and an int8 conversion wraps anything above 127.
+            output(i)%phi(k) = int((i_theta - 1) * cfg%n_phi_dir + i_phi, int16)
           end if
         end do
         k = nnode
